Migrate from Create React App (CRA) to Vite and enable TypeScript
- Replaced CRA setup with Vite
- Configured Vite to support TypeScript
- Updated project dependencies and scripts to align with Vite
- Refactored existing code to comply with TypeScript requirements
