+++ package.json
@@ -1,57 +1,38 @@
 {
   "name": "prog-compliance",
-  "version": "3.0.0",
   "private": true,
-  "dependencies": {
-    "@azure/arm-resourcegraph": "^4.2.1",
-    "@azure/identity": "^4.0.1",
-    "@azure/msal-browser": "^3.1.0",
-    "@azure/msal-react": "^2.0.3",
-    "@emotion/react": "^11.11.1",
-    "@emotion/styled": "^11.11.0",
-    "@fluentui/react": "^8.111.2",
-    "@fluentui/theme": "^2.6.35",
-    "@testing-library/jest-dom": "^5.17.0",
-    "@testing-library/react": "12.1.5",
-    "@testing-library/user-event": "^13.5.0",
-    "react": "^17.0.2",
-    "react-bootstrap": "^2.7.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^6.20.1",
-    "react-scripts": "^5.0.1",
-    "react-select": "^5.7.4",
-    "web-vitals": "^2.1.4"
+  "version": "3.0.0",
+  "type": "module",
+  "engines": {
+    "node": ">=20.0.0"
   },
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "jest",
-    "eject": "react-scripts eject"
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "lint": "eslint .",
+    "preview": "vite preview"
   },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+  "dependencies": {
+    "@azure/msal-browser": "^3.25.0",
+    "@azure/msal-react": "^2.1.1",
+    "@fluentui/react": "^8.121.2",
+    "@fluentui/react-hooks": "^8.8.14",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^6.26.2",
+    "vite-tsconfig-paths": "^5.0.1"
   },
   "devDependencies": {
-    "@babel/plugin-proposal-private-property-in-object": "^7.21.11"
-  },
-  "description": "This project was bootstrapped with [Create React App](https://github.com/facebook/create-react-app).",
-  "main": "index.js",
-  "keywords": [],
-  "author": "",
-  "license": "ISC"
+    "@eslint/js": "^9.11.1",
+    "@types/react": "^18.3.11",
+    "@types/react-dom": "^18.3.0",
+    "@vitejs/plugin-react": "^4.3.2",
+    "eslint": "^9.12.0",
+    "eslint-plugin-react-hooks": "^5.1.0-rc.0",
+    "eslint-plugin-react-refresh": "^0.4.12",
+    "globals": "^15.10.0",
+    "typescript": "^5.6.2",
+    "typescript-eslint": "^8.8.1",
+    "vite": "^5.4.8"
+  }
 }
