+++ src/main.tsx
@@ -1,10 +1,11 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './styles/index.css';
-import App from './App';
+// @ts-nocheck
 import { PublicClientApplication } from "@azure/msal-browser";
 import { MsalProvider } from "@azure/msal-react";
+import React from 'react';
+import ReactDOM from 'react-dom';
+import App from './App';
 import { msalConfig } from "./authConfig";
+import './styles/index.css';
 
 const msalInstance = new PublicClientApplication(msalConfig);
